% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colorpaletttes}
\alias{colorpaletttes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 20.
}
\usage{
colorpaletttes
}
\description{
Use \code{\link{chrisnolan_palettes}} to construct palettes of desired length.
}
\keyword{datasets}
