% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nolan_palettes}
\alias{nolan_palettes}
\title{A Christopher Nolan palette generator}
\usage{
nolan_palettes(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{DarkKnight}, \code{Inception1}, \code{Inception2},  \code{Inception3},
\code{Inception4}, \code{Tenet1},  \code{Tenet2},  \code{Interstellar1}, 
\code{Interstellar2}, \code{Interstellar3}, \code{Dunkirk1}, \code{Dunkirk2},
\code{Dunkirk3}, \code{Dunkirk4}, \code{Dunkirk5}, \code{Dunkirk6}, \code{Prestige1},
\code{Prestige2}, \code{Prestige3}, \code{Prestige4}}

\item{n}{Number of colors desired. Color schemes come from the website 
  \href{https://indiefilming.com/people/christopher-nolan/}{Indie Film} and others are from from the instagram page
\href{https://www.instagram.com/colorpalette.cinema/}{colorpalette.cinema}.
  If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Christopher Nolan movies. Currently, I only have a hanful of movies, but hope to include his entire filmography
}
\examples{
chrisnolan_palette("Tenet1")
chrisnolan_palette("Tenet2")
chrisnolan_palette("Interstellar1")
chrisnolan_palette("Interstellar2")
chrisnolan_palette("DarkKnight")


# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- chrisnolan_palette(21, name = "Tenet1", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
